tests/requirements.txt
attrs==18.1.0
coverage==4.5.1
flake8==3.5.0
isort==4.3.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.5.1
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
pytest-isort==0.1.0
pytest-mock==1.9.0
pytest-sugar==0.9.1
pytest-toolbox==0.4
